improved class not found
Before the class not found did not work well with multiple class
autoloaders already in existance, now it checks for them, and
re-arranges the class loaders so it works well.

PHP-Error is now reported as an internal file in the stack trace.

Also fixed a FireFox specific bug with the background height causing a
vertical scroll bar to appear.