code snippets are now flush with the left edge
Before code snippets could have a heavy amount of whitespace on the left
hand side,
this is now rectified.

If a code snippet is heavily indented, then it will be moved to be
closer to the left edge.

This removes large blocks of whitespace all down the left hand side.