fixed 5.3.0 issue, simplified class loading error code
Before checking for 'class not found' error'd on 5.3.0,
this is because it still retained the original code, where
it jumbled around class loaders (and some of the functions
don't work correctly with 5.3.0).

Since PHP Error moved over to recording class not found,
instead of raising an error, this code could just be entirely
removed.