#9 remains silent if returning non-HTML
Sometimes people request JS and CSS through PHP scripts,
in a non-ajaxy way. For example linking to a PHP script in a script tag.

This change will make PHP Error be silent if it believes it is on one of
those occasions.

It also changes the 'getResponseHeaders' method, to merge the apache
and non-apache headers together. It seems one can contain info the other
does not.