#29 fixed not showing up in Wordpress
There are two issues that have been fixed:

* the content-type used by wordpress was unknown,
because it included the character set. For example:

text/html; utf-8;

So new matching is used to just check if an allowed
type is present, and doesn't care if extra content is
there (such as 'UTF-8').

* Second problem is that the page flushes before
all sections are outputted. As a result, errors within
a section, do not replace the whole page.
The fix is to buffer all output, by default, and then
output at the end. This is now added.