#29 added IIS support, thinner font, backgrund now fixed
* IIS support is added. When FastCGI logs an error,
IIS take over the request by default, and then show it.
This meant PHP Error was failing when it logged an error.
Now we don't log an error if PHP Error is running (since
we are just going to tell them)

* background div is now fixed, to cover up anything that
has outputted before the error is displayed.

* The font set in a previous commit was incorrect,
it's now set to the correct, thinner version.